+++ KSP2Runtime/KSPOrbit/KSPOrbitModule.Orbit.cs
@@ -177,7 +177,7 @@ namespace KontrolSystem.KSP.Runtime.KSPOrbit {
             [KSMethod(Description = @"Gives the true anomaly (in a's orbit) at which a crosses its descending node with b's orbit.
                 The returned value is always between 0 and 360.")]
             double DescendingNodeTrueAnomaly(IOrbit b);
-            
+
             [KSMethod(Description = @"Returns the next time at which a will cross its ascending node with b.
                 For elliptical orbits this is a time between UT and UT + a.period.
                 For hyperbolic orbits this can be any time, including a time in the past if
@@ -204,7 +204,7 @@ namespace KontrolSystem.KSP.Runtime.KSPOrbit {
                 occurs at a true anomaly that a does not actually ever attain.
                 If `ut` is omitted the current time will be used")]
             double TimeOfDescendingNode(IOrbit b, Option<double> maybeUt = new Option<double>());
-            
+
             [KSField(Description = "Get the relative position of the ascending node.")]
             Vector3d RelativeAscendingNode { get; }