+++ KSP2Runtime/Testing/MockBody.cs
@@ -118,8 +118,8 @@ namespace KontrolSystem.KSP.Runtime.Testing {
         public KSPOrbitModule.IOrbit Orbit => orbit;
 
         public Option<KSPOrbitModule.IBody> AsBody => new Option<KSPOrbitModule.IBody>(this);
-            
-        public Option<KSPVesselModule.VesselAdapter> AsVessel  => new Option<KSPVesselModule.VesselAdapter>();
+
+        public Option<KSPVesselModule.VesselAdapter> AsVessel => new Option<KSPVesselModule.VesselAdapter>();
 
         public Option<KSPVesselModule.ModuleDockingNodeAdapter> AsDockingPort => new Option<KSPVesselModule.ModuleDockingNodeAdapter>();